Add lincheck and fix some linearization problems.
