+++ csc/2019/2.LockFreeSet/GoncharovAV/src/main/java/org/sample/jcstress_tests/RemoveTest.java
@@ -28,60 +28,72 @@
  * ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF
  * THE POSSIBILITY OF SUCH DAMAGE.
  */
-package org.sample.test;
+package org.sample.jcstress_tests;
 
 import org.openjdk.jcstress.annotations.*;
 import org.openjdk.jcstress.infra.results.*;
 import org.sample.LockFreeSet;
 import org.sample.LockFreeSetImpl;
 
+import java.util.HashSet;
 import java.util.Random;
 
 // See jcstress-samples or existing tests for API introduction and testing guidelines
 
 @JCStressTest
 // Outline the outcomes here. The default outcome is provided, you need to remove it:
-@Outcome(id = "0, 0", expect = Expect.ACCEPTABLE, desc = "Find all values.")
-@Outcome(id = "0, 1", expect = Expect.FORBIDDEN, desc = "Some values was not found.")
+@Outcome(id = "0", expect = Expect.ACCEPTABLE, desc = "All values were successfully removed.")
+@Outcome(id = "1", expect = Expect.FORBIDDEN, desc = "Some values was not removed.")
 @State
-public class ContainsTest {
+public class RemoveTest {
 
     private final int maxValue = 10000;
     private final int nElements = 1000;
 
     private final LockFreeSet<Integer> set = new LockFreeSetImpl<>();
 
-    private final int[] add_data;
-    private final int[] contains_data;
+    private final HashSet<Integer> add_data = new HashSet<>();
+    private final HashSet<Integer> remove_data = new HashSet<>();
 
-    ContainsTest(){
-        add_data = new int[nElements];
-        contains_data = new int[nElements];
-
+    RemoveTest() {
         Random r = new Random();
 
-        for (int i=0; i< nElements; ++i)
-            add_data[i] = r.nextInt(maxValue) + 1;
+        for (int i = 0; i < nElements; ++i)
+            add_data.add(r.nextInt(maxValue) + 1);
 
-        for (int i=0; i< nElements; ++i) {
-            contains_data[i] = r.nextInt(maxValue) + 1;
-            set.add(contains_data[i]);
+        for (int i = 0; i < nElements; ++i) {
+            int value = r.nextInt(maxValue) + 1;
+            if (!add_data.contains(value)) {
+                remove_data.add(value);
+                set.add(value);
+            }
         }
 
     }
 
     @Actor
-    public void actor1(II_Result r) {
-        for (int value:add_data){
+    public void actor1() {
+        for (int value : add_data) {
             set.add(value);
         }
     }
 
     @Actor
-    public void actor2(II_Result r) {
+    public void actor2() {
+        for (int value : remove_data) {
+            set.remove(value);
+        }
+    }
+
+    @Arbiter
+    public void arbiter(I_Result r) {
+
         boolean ok = true;
-        for (int value:contains_data){
+        for (int value : add_data) {
             ok &= set.contains(value);
+        }
+        for (int value : remove_data) {
+            ok &= !set.contains(value);
         }
 
         if (ok)
@@ -89,5 +101,5 @@ public class ContainsTest {
         else
             r.r1 = 1;
     }
-}
 
+}
